chore: create generator for it
