+++ playwright/async_api/_generated.py
@@ -9756,7 +9756,7 @@ class Playwright(AsyncBase):
         super().__init__(obj)
 
     @property
-    def devices(self) -> Devices:
+    def devices(self) -> "Devices":
         """Playwright.devices
 
         Returns a dictionary of devices to be used with `browser.new_context()` or `browser.new_page()`.
@@ -9783,7 +9783,7 @@ class Playwright(AsyncBase):
 
         Returns
         -------
-        {Blackberry PlayBook: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Blackberry PlayBook landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, BlackBerry Z30: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, BlackBerry Z30 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Note 3: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Note 3 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Note II: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Note II landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S III: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S III landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S5: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S5 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S8: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S8 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S9+: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy S9+ landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Tab S4: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Galaxy Tab S4 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad (gen 6): {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad (gen 6) landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad (gen 7): {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad (gen 7) landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad Mini: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad Mini landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad Pro 11: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPad Pro 11 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 6: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 6 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 6 Plus: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 6 Plus landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 7: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 7 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 7 Plus: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 7 Plus landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 8: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 8 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 8 Plus: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 8 Plus landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone SE: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone SE landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone X: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone X landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone XR: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone XR landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11 Pro: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11 Pro landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11 Pro Max: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 11 Pro Max landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12 Pro: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12 Pro landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12 Pro Max: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, iPhone 12 Pro Max landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, JioPhone 2: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, JioPhone 2 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Kindle Fire HDX: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Kindle Fire HDX landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, LG Optimus L70: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, LG Optimus L70 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Microsoft Lumia 550: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Microsoft Lumia 550 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Microsoft Lumia 950: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Microsoft Lumia 950 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 10: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 10 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 4: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 4 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 5: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 5 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 5X: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 5X landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 6: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 6 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 6P: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 6P landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 7: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nexus 7 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nokia Lumia 520: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nokia Lumia 520 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nokia N9: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Nokia N9 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 2: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 2 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 2 XL: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 2 XL landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 3: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 3 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 4: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 4 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 4a (5G): {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 4a (5G) landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 5: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}, Pixel 5 landscape: {viewport: {width: int, height: int}, user_agent: str, device_scale_factor: str, is_mobile: bool, has_touch: bool}}
+        Devices
         """
         return mapping.from_impl(self._impl_obj.devices)
